[REF] fs_storage: refactor test class to make it reusable
