+++ fs_storage/tests/__init__.py
@@ -1,1 +1,2 @@
+from . import common
 from . import test_fs_storage
